backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
transformers>=4.50.0
accelerate>=0.26.0
soundfile>=0.13.0
librosa>=0.10.0
nemo_toolkit[asr]>=2.0.0
# ONNX ASR (lightweight Parakeet)
onnx-asr>=0.5.0
onnxruntime-gpu>=1.20.0
optimum>=1.24.0
# torch: provided by base Docker image
